backend_py/requirements.txt
# ===============================

# Nexus – Literature Review Engine

# CPU-only (no GPU)

# ===============================

# --- Core API ---

fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
requests>=2.31
httpx==0.28.1
python-multipart>=0.0.9
python-dotenv

# --- Google GenAI ---

google-genai

# --- Vector DB ---

qdrant-client>=1.9
numpy>=1.24.0

# --- NLP / Embeddings ---

sentence-transformers
transformers==4.51.3
nltk==3.9.1

# --- File Parsing ---

pypdf>=4.0
PyMuPDF>=1.24
python-docx>=1.1
openpyxl>=3.1
beautifulsoup4

# --- Report Generation ---

fpdf2
